requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
